## Proposed changes

This changes a few problems I observed in our memory allocator.

1. In some cases the cache is really full but nothing fits. Instead of trying to clear the cache, the allocator waits and in some extreme cases even fails when there is memory technically "available" but it's held by the cache. (If there are no scheduled jobs, nothing in the cache fits, we don't clear the cache, we just fail).

2. Change the cache fit policy to be a little stricter. It used to require the block to be less than twice the size. Now it requires `min(2 * size, size + vm_page_size)`. The `vm_page_size` is somewhat arbitrary but I think it's much better policy. Can save a lot of memory and when you get to big sizes, typically missing allocator the cache isn't such a huge deal as we fall back to allocating a new bugger.

3. Add a force swap parameter to the `malloc`. Once there are no more jobs to wait for, instead of just breaking the allocator will swap to the CPU. This makes it possible to do large model conversions for example.

4. Decrease the wait limit. I notice we often swap even at `1.5 * recommendedSize` (probably related to the amount of wired memory and other apps running). I cranked it down to just `recommendedSize`. That + the force allocate seems to work just as well in the cases I tested.

I will add some benchmarks before we decide to merge this. But want to solicit comments in parallel.
 
## Checklist

Put an `x` in the boxes that apply.

- [x] I have read the [CONTRIBUTING](https://github.com/ml-explore/mlx/blob/main/CONTRIBUTING.md) document
- [x] I have run `pre-commit run --all-files` to format my code / installed pre-commit prior to committing changes
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have updated the necessary documentation (if needed)