@awni 2023-12-29
### Before change

**Llama benchmark**
Time per layer per token: 3.9148330688476562 ms
Lower bound total time per token: 125.274658203125 ms

**Linear regression**
Loss 0.00005, |w-w*| = 0.00364, Throughput 3666.22979 (it/s)
Logistic regression
Loss 0.00005, |w-w*| = 0.00364, Throughput 3655.93535 (it/s)

**Mnist CPU**
Epoch 9: Test accuracy 0.948, Time 0.078 (s)

**Mnist GPU**
Epoch 9: Test accuracy 0.948, Time 0.168 (s)

**Transformer LM Batch Size 2**
Iter 20: Train loss 8.854, It/sec 2.291

**Transformer LM Batch Size 4**
Iter 20: Train loss 8.869, It/sec 1.198

**Transformer LM Batch Size 8**
Iter 20: Train loss 8.885, It/sec 0.599

### After change

**Llama benchmark**
Time per layer per token: 3.997039794921875 ms
Lower bound total time per token: 127.9052734375 ms

**Linear regression**
Loss 0.00005, |w-w*| = 0.00364, Throughput 3708.16439 (it/s)

**Logistic regression**
Loss 0.00005, |w-w*| = 0.00364, Throughput 3691.89505 (it/s)

**Mnist CPU**
Epoch 9: Test accuracy 0.948, Time 0.081 (s)

**Mnist GPU**
Epoch 9: Test accuracy 0.948, Time 0.167 (s)

**Transformer LM Batch Size 2**
Iter 20: Train loss 8.885, It/sec 2.290

**Transformer LM Batch Size 4**
Iter 20: Train loss 8.830, It/sec 1.194

**Transformer LM Batch Size 8**
Iter 20: Train loss 8.870, It/sec 0.530