@angeloskath 2023-12-30
Just the default :-) . On my M2 air it goes from a consistent 0.73 it/s for at least 60 iterations to ~0.63 it/s but it is also erratic in the sense that sometimes it starts low and slowly increases after several iterations or it starts ok at 0.72 and decreases. It has to do with the caching and waiting patterns...

When reducing the memory pressure by reducing the number of blocks it seems to perform the same. On the stable diffusion example it also seems to perform practically the same.