@awni 2023-12-30
> It has to do with the caching and waiting patterns...

Yes almost certainly. I changed a couple things there. In general it's going to be hard to find a setting that makes everything faster :\, but I will see if I can tune it a bit for the transformer case.