@awni 2024-01-02
@angeloskath I fixed one bug I introduced and partially reverted some of the GC strategy. I don't see an regression running the transformer benchmark with the default params on an 8GB machine (which is pretty high memory presssure). Could you pull and try running your benchmark again?